requirements.txt
agno
openai
python-dotenv
mcp
wikipedia
tavily-python
beautifulsoup4
lxml
requests
googlesearch-python
snowflake-id